Work on creating a data store for securities, and using its manager in TrendingFragment.

+++ TradeHero/TH_ANDROID/src/main/java/com/tradehero/th/persistence/TimelineManager.java
@@ -18,7 +18,7 @@ public class TimelineManager
     public List<TimelineItem> getTimeline(Query query, boolean forceReload) throws IOException
     {
         // TODO scope locking for current timeline of user
-        TimelineStore timelineStore = allTimelineStores.get().under(query.getId());
+        TimelineStore timelineStore = allTimelineStores.get().under((Integer) query.getId());
         timelineStore.setQuery(query);
         return forceReload ? dbCache.requestAndStore(timelineStore) : dbCache.loadOrRequest(timelineStore);
         // and unlock the scope
